fix: move tenderdash init to task factory
