+++ src/commands/setup.js
@@ -27,7 +27,6 @@ class SetupCommand extends BaseCommand {
    * @param {Object} args
    * @param {Object} flags
    * @param {DockerCompose} dockerCompose
-   * @param {Docker} docker
    * @param {generateBlsKeys} generateBlsKeys
    * @param {ConfigCollection} configCollection
    * @param {initializeTenderdashNode} initializeTenderdashNode
@@ -37,6 +36,7 @@ class SetupCommand extends BaseCommand {
    * @param {writeServiceConfigs} writeServiceConfigs
    * @param {startNodeTask} startNodeTask
    * @param {platformInitTask} platformInitTask
+   * @param {tenderdashInitTask} tenderdashInitTask
    * @return {Promise<void>}
    */
   async runWithDependencies(
@@ -53,7 +53,6 @@ class SetupCommand extends BaseCommand {
       verbose: isVerbose,
     },
     dockerCompose,
-    docker,
     generateBlsKeys,
     configCollection,
     initializeTenderdashNode,
@@ -63,6 +62,7 @@ class SetupCommand extends BaseCommand {
     writeServiceConfigs,
     startNodeTask,
     platformInitTask,
+    tenderdashInitTask,
   ) {
     let config;
 
@@ -197,35 +197,7 @@ class SetupCommand extends BaseCommand {
       },
       {
         title: 'Initialize Tenderdash',
-        task: async (ctx) => {
-          const isValidatorKeyEmpty = Object.keys(config.get('platform.drive.tenderdash.validatorKey')).length === 0;
-          const isNodeKeyEmpty = Object.keys(config.get('platform.drive.tenderdash.nodeKey')).length === 0;
-          const isGenesisEmpty = Object.keys(config.get('platform.drive.tenderdash.genesis')).length === 0;
-
-          const { Volumes: existingVolumes } = await docker.listVolumes();
-          const { COMPOSE_PROJECT_NAME: composeProjectName } = config.toEnvs();
-          const isDataVolumeMissing = !existingVolumes.find((v) => v.Name === `${composeProjectName}_drive_tenderdash`);
-
-          if (isValidatorKeyEmpty || isNodeKeyEmpty || isNodeKeyEmpty || isDataVolumeMissing) {
-            const [validatorKey, nodeKey, genesis] = await initializeTenderdashNode(config);
-
-            if (isValidatorKeyEmpty) {
-              config.set('platform.drive.tenderdash.validatorKey', validatorKey);
-            }
-
-            if (isNodeKeyEmpty) {
-              config.set('platform.drive.tenderdash.nodeKey', nodeKey);
-            }
-
-            if (isGenesisEmpty) {
-              if (ctx.preset === PRESET_LOCAL) {
-                genesis.initial_core_chain_locked_height = 1000;
-              }
-
-              config.set('platform.drive.tenderdash.genesis', genesis);
-            }
-          }
-        },
+        task: () => tenderdashInitTask(config),
       },
       {
         title: 'Update config',
