+++ src/createDIContainer.js
@@ -45,6 +45,7 @@ const registerMasternode = require('./core/wallet/registerMasternode');
 const generateToAddressTaskFactory = require('./listr/tasks/wallet/generateToAddressTaskFactory');
 const registerMasternodeTaskFactory = require('./listr/tasks/registerMasternodeTaskFactory');
 const platformInitTaskFactory = require('./listr/tasks/platform/initTaskFactory');
+const tenderdashInitTaskFactory = require('./listr/tasks/tenderdash/initTaskFactory');
 const startNodeTaskFactory = require('./listr/tasks/startNodeTaskFactory');
 
 const createTenderdashRpcClient = require('./tenderdash/createTenderdashRpcClient');
@@ -136,6 +137,7 @@ async function createDIContainer(options) {
     generateToAddressTask: asFunction(generateToAddressTaskFactory),
     registerMasternodeTask: asFunction(registerMasternodeTaskFactory),
     platformInitTask: asFunction(platformInitTaskFactory),
+    tenderdashInitTask: asFunction(tenderdashInitTaskFactory),
     startNodeTask: asFunction(startNodeTaskFactory),
   });
